Move formatting to LSP store
